requirements.txt
certifi==2019.3.9
Django==2.2.1
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.9.0
mysqlclient==1.4.2.post1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
wincertstore==0.2
